Add new Code

+++ src/IronCenter.Service/Domain/Employes/Employee.cs
@@ -7,7 +7,7 @@ namespace IronCenter.Service.Domain.Employes
     {
         public string FullName { get; set; } = string.Empty;
         public string PhoneNumber { get; set; } = string.Empty ;
-        public int PositionId {  get; set; }
+        public long PositionId {  get; set; }
         public string PositionName { get; set; } = string.Empty;
         public decimal DailySalary { get; set; }
         public List<Salary> Salaries { get; set; }
